changed all config files and rollup compiles
