+++ lib/SailsCalls.js
@@ -1467,4 +1467,3 @@ class SailsCalls {
 }
 
 export { SailsCalls };
-//# sourceMappingURL=index.esm.js.map
