+++ lib/cjs/SailsCalls.cjs
@@ -1,5 +1,3 @@
-'use strict';
-
 var sailsJs = require('sails-js');
 var sailsJsParser = require('sails-js-parser');
 var api = require('@gear-js/api');
@@ -1469,4 +1467,3 @@ class SailsCalls {
 }
 
 exports.SailsCalls = SailsCalls;
-//# sourceMappingURL=index.cjs.js.map
