+++ lib/cjs/index.cjs
@@ -1,0 +1,5 @@
+var SailsCalls = require('./SailsCalls.cjs');
+
+
+
+exports.SailsCalls = SailsCalls.SailsCalls;
