+++ lib/index.d.ts
@@ -1,0 +1,2 @@
+export type { GasLimitType, ContractAddress, SailsInstance, SailsCallsContractsData, ContractData, NewContractData, ICreateVoucher, IBasicUpdateVoucherData, ITokensToAddToVoucher, IRenewVoucherAmountOfBlocks, ISailsCalls, SponsorData, SailsCallsError, ISailsQueryOptions, ISailsCommandOptions, ICommandResponse, WalletSigner, SailsCallbacks, CallbackType, AccountSigner } from './types';
+export { SailsCalls } from './SailsCalls';
