+++ lib/index.js
@@ -1,0 +1,1 @@
+export { SailsCalls } from './SailsCalls.js';
