+++ tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+      "declaration": true
+    },
+    "include": ["src"]
+  }
