+++ tsconfig.cjs.json
@@ -1,0 +1,6 @@
+{
+    "extends": "./tsconfig.build.json",
+    "compilerOptions": {
+      "declaration": false
+    }
+  }
